Refactor keyboard_handler into input_apis module
